Moved the ResultState from domain to Core
