+++ domain/build.gradle.kts
@@ -38,6 +38,7 @@ android {
 
 
 dependencies {
+    implementation(project(":core"))
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
