+++ domain/src/main/java/com/practical/domain/repository/CharacterRepository.kt
@@ -3,7 +3,7 @@ package com.practical.domain.repository
 import com.practical.domain.CharacterModel
 import com.practical.domain.CharactersListModel
 import com.practical.domain.EpisodeModelDetails
-import com.practical.domain.ResultState
+import com.practical.core.ResultState
 import kotlinx.coroutines.flow.Flow
