+++ domain/src/main/java/com/practical/domain/usecases/GetCharacterUseCase.kt
@@ -1,12 +1,12 @@
 package com.practical.domain.usecases
 
 import com.practical.domain.CharacterModel
-import com.practical.domain.ResultState
+import com.practical.core.ResultState
 import com.practical.domain.repository.CharacterRepository
 import kotlinx.coroutines.flow.Flow
 
 class GetCharacterUseCase(private val repository: CharacterRepository) {
-     operator fun invoke(id: String): Flow<ResultState<CharacterModel>> {
+     operator fun invoke(id: String): Flow<com.practical.core.ResultState<CharacterModel>> {
         return repository.getCharacter(id)
     }
 }
