+++ domain/src/main/java/com/practical/domain/usecases/GetEpisodeUseCases.kt
@@ -1,12 +1,12 @@
 package com.practical.domain.usecases
 
 import com.practical.domain.EpisodeModelDetails
-import com.practical.domain.ResultState
+import com.practical.core.ResultState
 import com.practical.domain.repository.CharacterRepository
 import kotlinx.coroutines.flow.Flow
 
 class GetEpisodeUseCases(private val repository: CharacterRepository) {
-    operator fun invoke(id: String): Flow<ResultState<EpisodeModelDetails>> {
+    operator fun invoke(id: String): Flow<com.practical.core.ResultState<EpisodeModelDetails>> {
         return repository.getEpisodeId(id)
     }
 }
