+++ domain/src/test/java/com/practical/domain/GetCharacterUseCaseTest.kt
@@ -21,15 +21,15 @@ class GetCharacterUseCaseTest {
         // Given
         val characterId = "123"
         coEvery { characterRepository.getCharacter(characterId) } returns flow {
-            emit(ResultState.Loading)
-            emit(ResultState.Success(CharacterModel(id = characterId, name = "Test Character")))
+            emit(com.practical.core.ResultState.Loading)
+            emit(com.practical.core.ResultState.Success(CharacterModel(id = characterId, name = "Test Character")))
         }
 
         // When
         getCharacterUseCase(characterId).test {
             // Collecting the result emitted by the flow
-            assertTrue(awaitItem() is ResultState.Loading)
-            assertTrue(awaitItem() is ResultState.Success)
+            assertTrue(awaitItem() is com.practical.core.ResultState.Loading)
+            assertTrue(awaitItem() is com.practical.core.ResultState.Success)
             awaitComplete() // Then
         }
     }
@@ -40,15 +40,15 @@ class GetCharacterUseCaseTest {
         val characterId = "123"
         val characterModel = CharacterModel(id = characterId, name = "Test Character")
         coEvery { characterRepository.getCharacter(characterId) } returns flow {
-            emit(ResultState.Success(characterModel))
+            emit(com.practical.core.ResultState.Success(characterModel))
         }
 
         // When
         getCharacterUseCase(characterId).test {
             // Collecting the result emitted by the flow
             val result = awaitItem()
-            assertTrue(result is ResultState.Success)
-            assertEquals(characterModel, (result as ResultState.Success).data)
+            assertTrue(result is com.practical.core.ResultState.Success)
+            assertEquals(characterModel, (result as com.practical.core.ResultState.Success).data)
             awaitComplete() // Then
         }
     }
@@ -58,15 +58,15 @@ class GetCharacterUseCaseTest {
         // Given
         val characterId = "123"
         coEvery { characterRepository.getCharacter(characterId) } returns flow {
-            emit(ResultState.Error(RuntimeException("Character not found")))
+            emit(com.practical.core.ResultState.Error(RuntimeException("Character not found")))
         }
 
         // When
         getCharacterUseCase(characterId).test {
             // Collecting the result emitted by the flow
             val result = awaitItem()
-            assertTrue(result is ResultState.Error)
-            assertEquals("Character not found", (result as ResultState.Error).exception.message)
+            assertTrue(result is com.practical.core.ResultState.Error)
+            assertEquals("Character not found", (result as com.practical.core.ResultState.Error).exception.message)
             awaitComplete() // Then
         }
     }
@@ -77,15 +77,15 @@ class GetCharacterUseCaseTest {
         // Given
         val characterId = "123"
         coEvery { characterRepository.getCharacter(characterId) } returns flow {
-            emit(ResultState.Error(RuntimeException("Something went wrong")))
+            emit(com.practical.core.ResultState.Error(RuntimeException("Something went wrong")))
         }
 
         //When
         getCharacterUseCase(characterId).test {
             // Collecting the result emitted by the flow
             val result = awaitItem()
-            assertTrue(result is ResultState.Error)
-            assertEquals("Something went wrong", (result as ResultState.Error).exception.message)
+            assertTrue(result is com.practical.core.ResultState.Error)
+            assertEquals("Something went wrong", (result as com.practical.core.ResultState.Error).exception.message)
             awaitComplete() //Then
         }
     }
