+++ domain/src/test/java/com/practical/domain/GetCharactersUseCaseTest.kt
@@ -20,10 +20,10 @@ class GetCharactersUseCaseTest {
     fun `should return Loading state when fetching characters`() = runTest {
         //Given
         coEvery { characterRepository.getCharactersList() } returns flow {
-            emit(ResultState.Loading)
+            emit(com.practical.core.ResultState.Loading)
             //When
             emit(
-                ResultState.Success(
+                com.practical.core.ResultState.Success(
                     listOf(
                         CharactersListModel(id = "1", name = "Character One")
                     )
@@ -34,10 +34,10 @@ class GetCharactersUseCaseTest {
         //Then
         getCharactersUseCase().test {
             // Collecting the first result (Loading)
-            assertTrue(awaitItem() is ResultState.Loading)
+            assertTrue(awaitItem() is com.practical.core.ResultState.Loading)
 
             // Collecting the second result (Success)
-            assertTrue(awaitItem() is ResultState.Success)
+            assertTrue(awaitItem() is com.practical.core.ResultState.Success)
             awaitComplete() // Ensure the flow completes
         }
     }
@@ -52,15 +52,15 @@ class GetCharactersUseCaseTest {
         )
 
         coEvery { characterRepository.getCharactersList() } returns flow {
-            emit(ResultState.Success(characterList))
+            emit(com.practical.core.ResultState.Success(characterList))
         }
 
 
         getCharactersUseCase().test {
             // Collecting the result emitted by the flow
             val result = awaitItem()
-            assertTrue(result is ResultState.Success)
-            assertEquals(characterList, (result as ResultState.Success).data)
+            assertTrue(result is com.practical.core.ResultState.Success)
+            assertEquals(characterList, (result as com.practical.core.ResultState.Success).data)
             awaitComplete() // Ensure the flow completes
         }
     }
@@ -70,17 +70,17 @@ class GetCharactersUseCaseTest {
         //Given
         val exception = RuntimeException("Error fetching characters")
         coEvery { characterRepository.getCharactersList() } returns flow {
-            emit(ResultState.Error(exception))
+            emit(com.practical.core.ResultState.Error(exception))
         }
 
         //When
         getCharactersUseCase().test {
             // Collecting the result emitted by the flow
             val result = awaitItem()
-            assertTrue(result is ResultState.Error)
+            assertTrue(result is com.practical.core.ResultState.Error)
             assertEquals(
                 "Error fetching characters",
-                (result as ResultState.Error).exception.message
+                (result as com.practical.core.ResultState.Error).exception.message
             )
             awaitComplete() //Then
         }
