+++ presentation/src/main/java/com/practical/presentation/screens/CharacterScreen.kt
@@ -27,7 +27,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
 import com.practical.domain.CharacterModel
 import com.practical.domain.EpisodeModel
-import com.practical.domain.ResultState
+import com.practical.core.ResultState
 import com.practical.presentation.R
 import com.practical.presentation.ui.theme.dimens
 import com.practical.presentation.viewmodel.CharacterDetailsViewModel
