+++ presentation/src/main/java/com/practical/presentation/screens/CharactersListScreen.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
 import com.practical.domain.CharactersListModel
-import com.practical.domain.ResultState
+import com.practical.core.ResultState
 import com.practical.presentation.R
 import com.practical.presentation.ui.theme.dimens
 import com.practical.presentation.viewmodel.CharacterViewModel
