+++ presentation/src/main/java/com/practical/presentation/viewmodel/CharacterViewModel.kt
@@ -2,18 +2,12 @@ package com.practical.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.practical.domain.CharacterModel
 import com.practical.domain.CharactersListModel
-import com.practical.domain.ResultState
-import com.practical.domain.usecases.GetCharacterUseCase
+import com.practical.core.ResultState
 import com.practical.domain.usecases.GetCharactersUseCase
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.launch
 import javax.inject.Inject
